Sort out extensions to PSI types
